% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toxic.R
\name{bandwidth}
\alias{bandwidth}
\title{Bandwidth toxic}
\usage{
bandwidth(rate = 0)
}
\arguments{
\item{rate}{rate in KB/s
'}
}
\description{
Limit a connection to a maximum number of kilobytes per second.
}

