% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toxic.R
\name{latency}
\alias{latency}
\title{Latency toxic}
\usage{
latency(latency = 0, jitter = 0)
}
\arguments{
\item{latency}{Latency, in milliseconds (1000 = 1s)}

\item{jitter}{Variation, in milliseconds}
}
\description{
Latency toxic.  Add a delay to all data going through the
proxy. The delay is equal to latency +/- jitter.
}

