% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toxic.R
\name{slicer}
\alias{slicer}
\title{Slicer toxic}
\usage{
slicer(average_size = 0, size_variation = 0, delay = 0)
}
\arguments{
\item{average_size}{size in bytes of an average packet}

\item{size_variation}{variation in bytes of an average packet
(should be smaller than average_size)}

\item{delay}{time in microseconds to delay each packet by}
}
\description{
Slices TCP data up into small bits, optionally adding a delay
between each sliced "packet".
}

