% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toxic.R
\name{slow_close}
\alias{slow_close}
\title{Delay TCP socket closure}
\usage{
slow_close(delay = 0)
}
\arguments{
\item{delay}{time in milliseconds}
}
\description{
Delay the TCP socket from closing until delay has elapsed.
}

