% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toxic.R
\name{timeout}
\alias{timeout}
\title{Timeout toxic}
\usage{
timeout(timeout = 0)
}
\arguments{
\item{timeout}{time in milliseconds}
}
\description{
Stops all data from getting through, and close the connection
after timeout. If timeout is 0, the connection won't close, and
data will be delayed until the toxic is disabled.
}

