% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxics.R
\name{toxic}
\alias{toxic}
\title{Toxics for use with toxiproxy}
\description{
Helper functions for creating toxics for use with toxiproxy.
These can be passed as the first argument to \code{$add} on a
toxic proxy in lieu of the \code{type} and \code{attributes}
arguments.
}
