% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toxic.R
\name{toxic_set}
\alias{toxic_set}
\title{Create a set of toxics}
\usage{
toxic_set(upstream = NULL, downstream = NULL)
}
\arguments{
\item{upstream}{Single toxic or list of toxics to apply to the
upstream connection}

\item{downstream}{Single toxic or list of toxics to apply to the
downstream connection}
}
\description{
Create a set of toxics to apply to a connection.
}
\details{
The possible toxics are \code{\link{latency}},
\code{\link{bandwidth}}, \code{\link{slow_close}},
\code{\link{timeout}} and \code{\link{slicer}}.
}
\examples{
toxic_set(timeout(1000))
# Simulate Australia:
toxic_set(upstream=list(bandwidth(10), latency(100)),
          downstream=list(bandwidth(50), latency(100)))
}

