% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxics.R
\name{latency}
\alias{latency}
\alias{bandwidth}
\alias{slow_close}
\alias{timeout}
\alias{slicer}
\alias{limit_data}
\title{The \code{latency} toxic adds a delay to all data going through
the proxy. The delay is equal to \code{latency +/- jitter}.}
\usage{
latency(latency, jitter = 0)

bandwidth(rate)

slow_close(delay)

timeout(timeout)

slicer(average_size, size_variation, delay = 0)

limit_data(bytes)
}
\arguments{
\item{latency}{Latency in milliseconds}

\item{jitter}{Time in milliseconds}

\item{rate}{Rate in KB per second}

\item{delay}{Time in milliseconds}

\item{timeout}{Time in milliseconds}

\item{average_size}{Size in bytes of an average packet}

\item{size_variation}{Variation in bytes of an average packet
(should be smaller than \code{average_size})}

\item{bytes}{The number of bytes it should transmit before
connection is closed}
}
\description{
The \code{latency} toxic adds a delay to all data going through
the proxy. The delay is equal to \code{latency +/- jitter}.

The \code{bandwidth} toxic limits a connection to a maximum number
of kilobytes per second.

The \code{slow_close} toxic delays the TCP socket from closing
until delay has elapsed.

The \code{timeout} toxic stops all data from getting through, and
closes the connection after \code{timeout}. If timeout is 0, the
connection won't close, and data will be delayed until the toxic
is removed.

The \code{slicer} toxic slices TCP data up into small bits,
optionally adding a delay between each sliced "packet".

The \code{limit_data} toxic closes the connection when transmitted
data exceeded limit.
}
