% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connection.R
\name{toxiproxy}
\alias{toxiproxy}
\title{Create a connection to the toxiproxy server}
\usage{
toxiproxy(host = "127.0.0.1", port = 8474)
}
\arguments{
\item{host}{Hostname to use; default assumes toxiproxy running locally}

\item{port}{Port to use; default assumes toxiproxy defaults}
}
\description{
Create a connection to the toxiproxy server.
}
\details{
The main entry point to the package.
}

