% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxiproxy_client.R
\name{toxiproxy_client}
\alias{toxiproxy_client}
\title{Make a toxiproxy client}
\usage{
toxiproxy_client(addr = NULL)
}
\arguments{
\item{addr}{The value address \emph{including protocol and port},
e.g., \code{http://toxiproxy.example.com:8474}.  If not given, the
default is the environment variable \code{TOXIPROXY_ADDR}.}
}
\description{
Make a toxiproxyy client.  This must be done before accessing the
toxiproxy server.  The default for arguments are controlled by
environment variables (see Details) and values provided as
arguments override these defaults.
}
\author{
Rich FitzJohn
}
