% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_install.R, R/server_manager.R
\name{toxiproxy_server_install}
\alias{toxiproxy_server_install}
\alias{toxiproxy_server}
\title{Control a test toxiproxy server}
\usage{
toxiproxy_server_install(quiet = FALSE, version = "v2.1.3")

toxiproxy_server(config = NULL, seed = NULL,
  if_disabled = testthat::skip)
}
\arguments{
\item{quiet}{Suppress progress bars on install}

\item{version}{Version of toxiproxy to install}

\item{config}{Not yet handled}

\item{seed}{Not yet handled}

\item{if_disabled}{Callback function to run if the toxiproxy
server is not enabled.  The default, designed to be used within
tests, is \code{testthat::skip}.  Alternatively, inspect the
\code{$enabled} property of the returned object.}
}
\description{
Control a server for use with testing.  This is designed to be
used only by other packages that wish to run tests against a
toxiproxy server.  You will need to set
\code{TOXIPROXYR_SERVER_BIN_PATH} to point at the directory
containing the toxiproxy binary.
}
\details{
The function \code{toxiproxy_server_install} will install a
server, but \emph{only} if the user opts in by setting the
environment variable \code{TOXIPROXYR_SERVER_INSTALL} to
\code{"true"}, and by setting \code{TOXIPROXYR_SERVER_BIN_PATH} to
the directory where the binary should be downloaded to.  This will
download a ~10MB binary from
\url{https://github.com/Shopify/toxiproxy/releases} so use with
care.  It is intended \emph{only} for use in automated testing
environments.
}
